provider/requirements.txt
attrs
fastapi
fuzzywuzzy
kubernetes
loguru
pydantic
pymongo
pyyaml
uvicorn